+++ src/main/java/org/acl/deepspark/utils/MnistLoader.java
@@ -23,30 +23,30 @@ public class MnistLoader implements Serializable {
 
 	public static Sample[] loadIntoSamples(String path, boolean normalize) {
 		System.out.println("Data Loading...");
-		double label;
+		float label;
 		int[] dimData = {1, dimRows, dimRows};
 		BufferedReader reader = null;
 		ArrayList<Sample> samples = new ArrayList<Sample>();
 		
-		double[] featureVec = new double[dimRows * dimRows];
+		float[] featureVec = new float[dimRows * dimRows];
 		try {
 			reader = new BufferedReader(new FileReader(path));
 			String line = null;
 			String[] feature = null;
 			while((line = reader.readLine()) != null) {
 				feature = line.split("\t");
-				double[] labelVec = new double[dimLabel];
-				label = Double.parseDouble(feature[dimRows * dimRows]);
+				float[] labelVec = new float[dimLabel];
+				label = Float.parseFloat(feature[dimRows * dimRows]);
 				for(int i = 0; i < dimLabel; i++) {
-					labelVec[i] = (label == i) ?  1.0 : 0.0;
+					labelVec[i] = (label == i) ?  1 : 0;
 				}
 				for(int i = 0; i < feature.length - 1;i++)
-					featureVec[i] = Double.parseDouble(feature[i]);
+					featureVec[i] = Float.parseFloat(feature[i]);
 				
 				Sample s = new Sample();
 				s.data = Tensor.create(featureVec, dimData).transpose();
 				if (normalize)
-					s.data.subi(128.0).divi(128.0);
+					s.data.subi(128).divi(128);
 				s.label = Tensor.create(labelVec, new int[] {dimLabel});
 
 				samples.add(s);
@@ -80,22 +80,22 @@ public class MnistLoader implements Serializable {
 			@Override
 			public Sample call(String v1) throws Exception {
 				String[] feature = v1.split("\t");
-				double label = Double.parseDouble(feature[dimRows * dimRows]);
-				double[] featureVec = new double[dimRows * dimRows];
-				double[] labelVec = new double[dimLabel];
+				float label = Float.parseFloat(feature[dimRows * dimRows]);
+				float[] featureVec = new float[dimRows * dimRows];
+				float[] labelVec = new float[dimLabel];
 				for (int i = 0; i < dimLabel; i++) {
-					labelVec[i] = (label == i) ? 1.0 : 0.0;
+					labelVec[i] = (label == i) ? 1 : 0;
 				}
 
 				for (int i = 0; i < feature.length -1; i++)
-					featureVec[i] = Double.parseDouble(feature[i]);
+					featureVec[i] = Float.parseFloat(feature[i]);
 				
 				Sample s = new Sample();
 				int[] dimData = {1, dimRows, dimRows};
 
 				s.data = Tensor.create(featureVec, dimData);
 				if (normalize)
-					s.data.subi(128.0).divi(128.0);
+					s.data.subi(128).divi(128);
 				s.label = Tensor.create(labelVec, new int[] {dimLabel});
 
 				return s;
@@ -106,11 +106,11 @@ public class MnistLoader implements Serializable {
 
 	public static Sample[] loadFromHDFS(String path, boolean normalize) {
 		System.out.println("Data Loading...");
-		double label;
+		float label;
 		int[] dimData = {1, dimRows, dimRows};
 		BufferedReader reader = null;
 		ArrayList<Sample> samples = new ArrayList<Sample>();
-		double[] featureVec = new double[dimRows * dimRows];
+		float[] featureVec = new float[dimRows * dimRows];
 
 		try {
 			Path p = new Path(path);
@@ -121,19 +121,19 @@ public class MnistLoader implements Serializable {
 
 			while ((line = reader.readLine()) != null) {
 				feature = line.split("\t");
-				double[] labelVec = new double[dimLabel];
-				label = Double.parseDouble(feature[dimRows * dimRows]);
+				float[] labelVec = new float[dimLabel];
+				label = Float.parseFloat(feature[dimRows * dimRows]);
 				for (int i = 0; i < dimLabel; i++) {
-					labelVec[i] = (label == i) ? 1.0 : 0.0;
+					labelVec[i] = (label == i) ? 1 : 0;
 				}
 
 				for (int i = 0; i < feature.length -1; i++)
-					featureVec[i] = Double.parseDouble(feature[i]);
+					featureVec[i] = Float.parseFloat(feature[i]);
 
 				Sample s = new Sample();
 				s.data = Tensor.create(featureVec, dimData).transpose();
 				if (normalize)
-					s.data.subi(128.0).divi(128.0);
+					s.data.subi(128).divi(128);
 				s.label = Tensor.create(labelVec, new int[] {dimLabel});
 
 				samples.add(s);
